LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/nested-supercedes-common-to-both-axes-expected.txt

Box 1 Box 2 Box 3

FAIL scroller prefers nested area over area aligned in both axes. assert_equals: scroller followed box2 in y axis after layout change expected 100 but got 0
